adding RoomControler first version

+++ src/main/java/pl/vabank/game/access/RoomRepository.java
@@ -9,7 +9,7 @@ import pl.vabank.game.data.RoomData;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface RoomRepository extends JpaRepository<RoomData, Integer> {
+public interface RoomRepository extends JpaRepository<RoomData, Long> {
     // @Query("SELECT u FROM UserData u WHERE u.email = ?1")
     // public RoomData findByEmail(String email);
 }
